@@ -78,12 +78,23 @@
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<org.imixs.workflow.version>5.2.17</org.imixs.workflow.version>
-		<org.imixs.melman.version>1.0.22</org.imixs.melman.version>
+		<jakarta.version>10.0.0</jakarta.version>
+		<org.imixs.workflow.version>6.0.3-SNAPSHOT</org.imixs.workflow.version>
+		<org.imixs.melman.version>2.0.0</org.imixs.melman.version>
 		<apache.pdfbox.version>2.0.21</apache.pdfbox.version>
 		<microprofile.version>3.0</microprofile.version>
 		<microprofile-metrics.version>2.3</microprofile-metrics.version>
 	</properties>
+
+	<repositories>
+		<repository>
+			<id>snapshots-repo</id>
+			<url>https://oss.sonatype.org/content/repositories/snapshots</url>
+			<snapshots>
+				<enabled>true</enabled>
+			</snapshots>
+		</repository>
+	</repositories>
 
 	<build>
 		<plugins>
@@ -94,8 +105,8 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.8.1</version>
 				<configuration>
-					<source>1.8</source>
-					<target>1.8</target>
+					<source>11</source>
+					<target>11</target>
 				</configuration>
 			</plugin>
 			<!-- release management -->
@@ -284,18 +295,12 @@
 
 	<dependencies>
 
-		<!-- JDK 8 support -->
+		<!-- Jakarta EE -->
 		<dependency>
-		    <groupId>jakarta.xml.bind</groupId>
-		    <artifactId>jakarta.xml.bind-api</artifactId>
-		    <version>2.3.2</version>
+		    <groupId>jakarta.platform</groupId>
+		    <artifactId>jakarta.jakartaee-api</artifactId>
+		    <version>${jakarta.version}</version>
 		    <scope>provided</scope>
-		</dependency>		
-		<dependency>
-		    <groupId>org.glassfish.jaxb</groupId>
-		    <artifactId>jaxb-runtime</artifactId>
-		     <version>2.3.2</version>
-		     <scope>provided</scope>
 		</dependency>
 
 		<!-- Microprofile -->
@@ -306,14 +311,26 @@
 			<type>pom</type>
 			<scope>provided</scope>
 		</dependency>
-		<dependency>
-			<groupId>org.eclipse.microprofile.metrics</groupId>
-			<artifactId>microprofile-metrics-api</artifactId>
-			<version>${microprofile-metrics.version}</version>
-			<scope>provided</scope>
-		</dependency>
-
+	
 		<!-- Testing -->
+		<dependency>
+			<groupId>org.eclipse.parsson</groupId>
+			<artifactId>jakarta.json</artifactId>
+			<version>1.1.1</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+		    <groupId>org.glassfish.jaxb</groupId>
+		    <artifactId>jaxb-runtime</artifactId>
+		    <version>3.0.0</version>
+		    <scope>test</scope>
+		</dependency>
+    	<dependency>
+	       <groupId>jakarta.xml.bind</groupId>
+	       <artifactId>jakarta.xml.bind-api</artifactId>
+	       <version>3.0.0</version>
+	       <scope>test</scope>
+	    </dependency>		
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
