@@ -27,10 +27,10 @@
 
 package org.imixs.archive.service;
 
-import javax.annotation.PostConstruct;
-import javax.inject.Inject;
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
+import jakarta.annotation.PostConstruct;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
 
 import org.imixs.archive.service.resync.ResyncService;
 import org.imixs.archive.service.util.MessageService;
