@@ -2,8 +2,8 @@ package org.imixs.archive.service.ui;
 
 import java.io.Serializable;
 
-import javax.enterprise.context.SessionScoped;
-import javax.inject.Named;
+import jakarta.enterprise.context.SessionScoped;
+import jakarta.inject.Named;
 
 /**
  * Session Scoped CID Bean to hold the current error messages.
